requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
frozenlist==1.4.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
node==1.2.2
odict==1.9.0
openai==0.28.0
pip==24.2
plumber==1.7
python-dotenv==1.0.1
requests==2.32.3
setuptools==74.1.1
tqdm==4.66.5
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.9.8
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==7.0.3
zope.lifecycleevent==5.0
zope.proxy==5.3
